requirements.txt
asgiref==3.11.0
Django==6.0
django-ckeditor==6.7.3
django-cors-headers==4.6.0
django-js-asset==3.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
Pillow==11.1.0
PyJWT==2.10.1
sqlparse==0.5.4

# Production dependencies
gunicorn==23.0.0
psycopg2-binary==2.9.10
whitenoise==6.8.2
python-decouple==3.8
dj-database-url==2.3.0
django-cloudinary-storage==0.3.0
cloudinary==1.41.0
django-admin-sortable2==2.2.4
